requirements.txt
Babel==2.8.0
Django==2.2
django-autoslug==1.9.7
django-autoslug-field==0.2.3
django-phonenumber-field==4.0.0
phonenumbers==8.12.5
Pillow==7.1.2
pytz==2020.1
sqlparse==0.3.1
